Split up dataset app into subpackages
