+++ changelog/contact/split-up-dataset-into-subpackages.internal.rst
@@ -1,0 +1,1 @@
+Dataset app is split up into subpackages. i.e. datahub.dataset.views to datahub.dataset.contact.views
