+++ datahub/dataset/contact/pagination.py
@@ -1,0 +1,9 @@
+from rest_framework.pagination import CursorPagination
+
+
+class ContactsDatasetViewCursorPagination(CursorPagination):
+    """
+    Cursor Pagination for ContactsDatasetView
+    """
+
+    ordering = ('created_on', 'pk')
