+++ datahub/dataset/order/pagination.py
@@ -1,0 +1,9 @@
+from rest_framework.pagination import CursorPagination
+
+
+class OMISDatasetViewCursorPagination(CursorPagination):
+    """
+    Cursor Pagination for OMISDatasetView
+    """
+
+    ordering = ('created_on', 'pk')
